Extract jest-each type defs to global
